refactor(frontend): split course NFT card into Header/Image/Price/Actions subcomponents (#774)
